Update stand_alone_functions.R
@anastasia-mbithe good find, and good investigation. The issue here wasn't to do with the NA values, but instead that the function doesn't have a way to handle this "case".
The function, until now, has only handled lists. However, this is a data frame.

``` 
x <- rcorpora::corpora("religion/christian_saints")
head(x)
class(x)
```

So we just have to add in a new `if` statement for how to handle the data if it is a data frame (that is, in it's simplest case!).
I've done this and made the changes in this commit.
This is ready for you again now.