Update stand_alone_functions.R
@rdstern these should all work now. The main issue was with the religion one - it contains a list in a list in a list in a list in a list (`data$Abrahamic$Christianity$`Eastern Christianity`$`Russian Orthodox Old Believers``) so it took a little while.

I think there is a much nicer way to do it than I am doing. I have seen, for example, Danny call his own function in the function. I think that is what we would want here. I can spend some time on cleaning this up and looking into that, if you think it is worth it? Alternatively, I can do that on another branch once this is merged.